requirements.txt
boto3
python-dotenv
openai
requests
pillow
rembg
numpy
vertexai
pandas
torch
scikit-learn
joblib
transformers
torchvision
torchattacks